♻️refactor: 평점 타입 변경 - rating 타입 int에서 double로 변경

+++ backend/src/main/java/com/tripfriend/domain/review/dto/ReviewRequestDto.java
@@ -23,7 +23,6 @@ public class ReviewRequestDto {
 
     @Min(value = 1, message = "평점은 1점 이상이어야 합니다.")
     @Max(value = 5, message = "평점은 5점 이하여야 합니다.")
-    private int rating;
-
+    private double rating; // int -> double
     private Long placeId;
 }
